update the logic of changing user data

+++ src/pages/UserSettings/Settings/index.jsx
@@ -4,65 +4,63 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import './index.css';
 import { commentsSVG, createArticleSVG, deleteSVG, ratingSVG } from '../../../components/SvgSprite';
-import { logOut } from '../../../redux/slices/authorization.js';
 import { backendUrl } from '../../../variables.js';
+import { fetchChangeUserData } from '../../../redux/slices/authorization.js';
 
 function Settings() {
-  // create dispatch for redux
+  //dispatch for redux
   const dispatch = useDispatch();
   //user data from redux
-  const { fullName, email, avatarUrl, rating, userArticles, userComments } = useSelector(
+  const { avatarUrl, fullName, email, rating, userArticles, userComments } = useSelector(
     (state) => state.authorization.userData
   );
   //ref for avatar input
   const inputAvatar = React.useRef();
   //default avatar url
   const defaultAvatarUrl = 'http://localhost:3000/img/avatars/defaultAvatar.png';
-  //current avatar URL
-  const [currentAvatarUrl, setCurrentAvatarUrl] = React.useState('');
-  //readOnly status for input password because of autocompleate
-  const [readOnly, setReadOnly] = React.useState(true);
   //state for user data
   const [registrationData, setRegistrationData] = React.useState({
     avatarUrl,
     fullName,
     email,
     password: '',
-    confirmPassword: '',
-    currentPassword: ''
+    confirmPassword: ''
   });
-  //send registration data and get user data from backend
-  async function sendUserData() {
-    try {
-      const { data } = await axios.patch(`/authorization/changeData`, {
-        ...registrationData,
-        password: registrationData.password || registrationData.currentPassword
-      });
-      alert(data.message);
-      dispatch(logOut());
-    } catch (err) {
-      alert(err.response.data.message);
-    }
-  }
-  //check user data
-  function userDataValidation() {
+  //check user common data
+  function userCommonValidation() {
     if (registrationData.fullName.length < 2) {
       return "Перевірте ім'я";
     }
     if (!/^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/.test(registrationData.email)) {
       return 'Перевірте email';
     }
+    return 'Підтвердити';
+  }
+  //check user data
+  function userPasswordValidation() {
     if (
-      registrationData.password &&
-      (registrationData.password.length < 6 ||
-        registrationData.password !== registrationData.confirmPassword)
+      registrationData.password.length < 6 ||
+      registrationData.password !== registrationData.confirmPassword
     ) {
-      return 'Перевірте пароль';
-    }
-    if (!registrationData.currentPassword.length) {
-      return 'Вкажіть Ваш пароль';
+      return 'Перевірте новий пароль';
     }
     return 'Підтвердити';
+  }
+  //send common data and get user data from backend
+  async function sendNameData() {
+    dispatch(
+      fetchChangeUserData({
+        fullName: registrationData.fullName
+      })
+    );
+  }
+  //send password data and get user data from backend
+  async function sendPasswordData() {
+    dispatch(
+      fetchChangeUserData({
+        password: registrationData.password
+      })
+    );
   }
   //upload avatar to server
   async function uploadAvatar(file) {
@@ -74,30 +72,35 @@ function Settings() {
       const formData = new FormData();
       formData.append('image', file);
       const { data } = await axios.post(`/upload?dir=users`, formData);
-      currentAvatarUrl &&
-        (await axios.delete(
-          `delete/${currentAvatarUrl.slice(currentAvatarUrl.lastIndexOf('/') + 1)}?dir=users`
-        ));
-      setRegistrationData((prev) => ({
-        ...prev,
-        avatarUrl: backendUrl + data.url
-      }));
-      setCurrentAvatarUrl(data.url);
+      deleteAvatar();
+      setRegistrationData((prev) => ({ ...prev, avatarUrl: backendUrl + data.url }));
+      dispatch(
+        fetchChangeUserData({
+          avatarUrl: backendUrl + data.url
+        })
+      );
     } catch (err) {
       alert('Не вдалось завантажити аватар');
     }
   }
-  //delete avatar from server
-  function deleteAvatar() {
-    axios.delete(
-      `delete/${currentAvatarUrl.slice(currentAvatarUrl.lastIndexOf('/') + 1)}?dir=users`
-    );
-  }
   //clear avatar
   function clearAvatar() {
-    currentAvatarUrl && deleteAvatar();
+    deleteAvatar();
     inputAvatar.current.value = '';
     setRegistrationData((prev) => ({ ...prev, avatarUrl: defaultAvatarUrl }));
+    axios.patch(`/authorization/changeData`, {
+      avatarUrl: defaultAvatarUrl
+    });
+  }
+  //delete avatar from server
+  function deleteAvatar() {
+    if (registrationData.avatarUrl !== defaultAvatarUrl) {
+      axios.delete(
+        `delete/${registrationData.avatarUrl.slice(
+          registrationData.avatarUrl.lastIndexOf('/') + 1
+        )}?dir=users`
+      );
+    }
   }
   return (
     <div className="settings">
@@ -134,6 +137,10 @@ function Settings() {
             <button onClick={clearAvatar}>{deleteSVG}</button>
           )}
         </div>
+        <h3>{fullName}</h3>
+      </div>
+      <div className="settings__nameData">
+        <h3>Змінити ім'я:</h3>
         <div className="settings__inputField">
           <input
             type="text"
@@ -148,6 +155,7 @@ function Settings() {
         <div className="settings__inputField">
           <input
             type="email"
+            disabled
             placeholder=" "
             value={registrationData.email}
             onChange={(event) =>
@@ -156,14 +164,20 @@ function Settings() {
           />
           <div>E-mail</div>
         </div>
-        <h3>Зміна паролю:</h3>
+        <button
+          className="acceptButton"
+          onClick={sendNameData}
+          disabled={userCommonValidation() !== 'Підтвердити'}>
+          {userCommonValidation()}
+        </button>
+      </div>
+      <div className="settings__passwordData">
+        <h3>Змінити пароль:</h3>
         <div className="settings__inputField">
           <input
             type="password"
-            readOnly={readOnly}
-            onFocus={() => setReadOnly(false)}
-            onBlur={() => setReadOnly(true)}
             placeholder=" "
+            autoComplete="new-password"
             value={registrationData.password}
             onChange={(event) =>
               setRegistrationData((prev) => ({ ...prev, password: event.target.value }))
@@ -174,10 +188,8 @@ function Settings() {
         <div className="settings__inputField">
           <input
             type="password"
-            readOnly={readOnly}
-            onFocus={() => setReadOnly(false)}
-            onBlur={() => setReadOnly(true)}
             placeholder=" "
+            autoComplete="new-password"
             value={registrationData.confirmPassword}
             onChange={(event) =>
               setRegistrationData((prev) => ({ ...prev, confirmPassword: event.target.value }))
@@ -185,27 +197,11 @@ function Settings() {
           />
           <div>Підтвердіть пароль</div>
         </div>
-      </div>
-      <div className="settings__confirmField">
-        <div className="setting__advice">
-          Будь ласка переконайтесь, що зазначені дані відповідають дійсності.
-        </div>
-        <div className="settings__inputField">
-          <input
-            type="password"
-            placeholder=" "
-            value={registrationData.currentPassword}
-            onChange={(event) =>
-              setRegistrationData((prev) => ({ ...prev, currentPassword: event.target.value }))
-            }
-          />
-          <div>Ваш пароль</div>
-        </div>
         <button
           className="acceptButton"
-          onClick={sendUserData}
-          disabled={userDataValidation() !== 'Підтвердити'}>
-          {userDataValidation()}
+          onClick={sendPasswordData}
+          disabled={userPasswordValidation() !== 'Підтвердити'}>
+          {userPasswordValidation()}
         </button>
       </div>
     </div>
